hotfix(rename): rename and move the functions
