+++ main.py
@@ -3,14 +3,14 @@
 ## @Miou-zora PROJECT, 2023
 ## mirror-generator
 ## File description:
-## why do you look at this aweful code ...
+## main.py
 ##
 
 from github import Github
 from sys import argv
 import json
-from flag_friend.add_to_collaborators import *
-from flag_friend.argument_manager import *
+from add_to_collaborators import *
+from argument_manager import *
 from push_mirror import *
 from generate_mirror_workflow import *
 from generate_folders_with_repo import *
